% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRCoDa.R
\name{LRCoDa}
\alias{LRCoDa}
\alias{ilrregression}
\alias{robilrregression}
\title{Classical and robust regression of non-compositional (real) response on compositional and/or non-compositional predictors}
\usage{
LRCoDa(
  y,
  X,
  external = NULL,
  factor_column = NULL,
  method = "robust",
  method_pivot = "orthonormal"
)
}
\arguments{
\item{y}{The response which should be non-compositional (Better explanation)}

\item{X}{The compositional and/or non-compositional predictors as a matrix, data.frame or numeric vector (Probably a refactoring done)}

\item{external}{Specify the columns which are not part of the composition}

\item{factor_column}{Specify the column which includes factor levels}

\item{method}{If robust, the fast MM-type linear regression is applied, while with method
\dQuote{classical}, the conventional least squares regression is applied.}

\item{method_pivot}{Method to choose the pivot coordinates (parameters pivotvar and norm have then no effect)}
}
\value{
An object of class \sQuote{lmrob} or \sQuote{lm} and two summary
objects.
}
\description{
Delivers appropriate inference for regression of y on compositional and/or non-compositional input X.
}
\details{
Compositional explanatory variables should
}
\examples{

## How the content of sand of the agricultural and grazing land soils
## in Germany depend on relative contributions of the main chemical trace elements,
## their different soil types and the Annual mean temperature:
data("gemas")
gemas$COUNTRY <- as.factor(gemas$COUNTRY)
gemas_GER <- dplyr::filter(gemas, gemas$COUNTRY == 'GER')
y <- gemas_GER$sand
X <- dplyr::select(gemas_GER, c(MeanTemp, soilclass, Al:Zr))
LRCoDa(y, X, external = c('MeanTemp'), factor_column = 'soilclass', method='classical', method_pivot = 'orthonormal')
LRCoDa(y, X, external = c('MeanTemp'), factor_column = 'soilclass', method='robust', method_pivot = 'orthonormal')
}
\seealso{
\code{\link{lm}} and \code{\link{lmrob}}
}
\author{
Roman Wiedemeier
}
\keyword{compositional}
\keyword{data}
\keyword{linear}
\keyword{models}
\keyword{regression}
